fix: extract select users component
